Add new recipe maintainer
Hi there, I am adding myself as a maintainer to this package.

I work for Kitware in the ParaView team and I am personally in charge of performing the upstream releases for ParaView. I also maintain ParaView packages in other package managers (Spack/SUSE) and I am also the main maintainer of a multitude of conda-forge packages, many of them of software created or contributed by Kitware.

I want to be a maintainer here so that I can streamline releases and also issues to upstream.

Thanks!